+++ src/main/java/com/mongodb/kafka/connect/util/ConnectConfigException.java
@@ -23,8 +23,11 @@ public class ConnectConfigException extends ConfigException {
   private final String name;
   private final Object value;
 
+  private final String originalMessage;
+
   public ConnectConfigException(final String name, final Object value, final String message) {
     super(name, value, message);
+    this.originalMessage = message;
     this.name = name;
     this.value = value;
   }
@@ -35,5 +38,9 @@ public class ConnectConfigException extends ConfigException {
 
   public Object getValue() {
     return value;
+  }
+
+  public String getOriginalMessage() {
+    return originalMessage;
   }
 }